+++ boards/edgebadge/examples/ferris_img.rs
@@ -8,7 +8,7 @@
 #![no_std]
 #![no_main]
 
-use edgebadge::{self as hal, entry, pac, Pins};
+use edgebadge::{entry, hal, pac, Pins};
 use panic_halt as _;
 
 use embedded_graphics::image::Image;