+++ boards/edgebadge/examples/neopixel_adc_light.rs
@@ -6,11 +6,11 @@
 #![no_std]
 #![no_main]
 
-use edgebadge::{self as hal, entry, pac, Pins};
+use edgebadge::{entry, hal, pac, Pins};
 use panic_halt as _;
 
-use embedded_hal::digital::v1_compat::OldOutputPin;
 use hal::adc::Adc;
+use hal::ehal::digital::v1_compat::OldOutputPin;
 use hal::prelude::*;
 use hal::timer::SpinTimer;
 use hal::{clock::GenericClockController, delay::Delay};