+++ boards/pygamer/examples/neopixel_adc_battery.rs
@@ -8,7 +8,7 @@
 
 #[cfg(not(feature = "panic_led"))]
 use panic_halt as _;
-use pygamer::{self as hal, entry, pac, Pins};
+use pygamer::{entry, hal, pac, Pins};
 
 use hal::adc::Adc;
 use hal::pac::gclk::pchctrl::GEN_A::GCLK11;